refactor valuation tests into own class
